+++ packages/react-nft-checkout/src/components/AppButton/index.ts
@@ -1,0 +1,2 @@
+export type { AppButtonProps } from './AppButton'
+export { default } from './AppButton'