+++ packages/react-nft-checkout/src/theme/palette.theme.ts
@@ -1,0 +1,13 @@
+import { PaletteOptions } from '@mui/material'
+
+export const paletteTheme: PaletteOptions = {
+  secondary: {
+    main: '#99a0c0',
+  },
+  success: {
+    main: '#9ce680',
+  },
+  error: {
+    main: '#d32f2f',
+  },
+}